refactor : 컨트롤러 분기
